Refactor BuildFailure items into corresponding kt file
